refactor: moved _generated-pages.js into _generated directory
Just to keep things more consistent.
