+++ .changeset/kind-dogs-cross.md
@@ -1,0 +1,5 @@
+---
+'@waveplay/pilot': patch
+---
+
+refactor: moved \_generated-pages.js into \_generated directory
