+++ packages/pilot/src/client/ui/area.tsx
@@ -6,7 +6,7 @@ import { PilotRoute } from '../core/pilot'
 import { PilotRenderer } from './renderer'
 import type { PilotConfig, PilotRouteOptions, PilotStateProps } from '../types'
 import { usePilot } from '../core/use-pilot'
-import { importPage, pageRoutes } from '../../_generated-pages'
+import { importPage, pageRoutes } from '../../_generated/pages'
 
 interface PilotAreaProps {
 	autoLoad?: boolean
