+++ packages/pilot/tsup-gen.config.ts
@@ -1,7 +1,8 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-	entry: ['src/_generated', 'src/_generated-pages.ts'],
+	entry: ['src/_generated'],
+	outDir: 'dist/_generated',
 	format: ['cjs'],
 	bundle: false,
 	clean: false,
